% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savage_go_functions_v2.R
\name{get_entrez}
\alias{get_entrez}
\title{Get Entrez gene IDs from mouse IDs}
\usage{
get_entrez(gene_list)
}
\arguments{
\item{gene_list}{A character vector of mouse gene symbols}
}
\value{
A data frame with the original gene names and the associated Entrez IDs
}
\description{
Get Entrez gene IDs from mouse IDs
}
