% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savage_go_functions_v2.R
\name{get_human_id}
\alias{get_human_id}
\title{Get human gene names and Entrez IDs from mouse gene names}
\usage{
get_human_id(gene_list)
}
\arguments{
\item{gene_list}{A character vector of mouse gene symbols}
}
\value{
A data frame with the original mouse gene names and the associated human gene names and Entrez IDs
}
\description{
Get human gene names and Entrez IDs from mouse gene names
}
