% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savage_go_functions_v2.R
\name{go_gene_list}
\alias{go_gene_list}
\title{Run gene ontology on a list of mouse gene names}
\usage{
go_gene_list(gene_list, gene_df, go_type)
}
\arguments{
\item{gene_list}{A list of mouse gene names}

\item{gene_df}{A data frame with your mouse gene names and Entrez IDs}

\item{go_type}{The Gene Ontology type to run ("BP", "CC", or "MF")}
}
\value{
An enrichResult object
}
\description{
Run gene ontology on a list of mouse gene names
}
