% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savage_go_functions_v2.R
\name{go_reduce}
\alias{go_reduce}
\title{Removes redundancy in GO terms using rrvgo}
\usage{
go_reduce(go_ora, go_type)
}
\arguments{
\item{go_ora}{An enrichResult object}

\item{go_type}{The Gene Ontology type ("BP", "CC", or "MF")}
}
\value{
A data frame of reduced GO terms
}
\description{
Removes redundancy in GO terms using rrvgo
}
