% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savage_go_functions_v2.R
\name{go_reduced_plot}
\alias{go_reduced_plot}
\title{Creates a plot of the reduced GO terms}
\usage{
go_reduced_plot(
  go_ora,
  reducedTerms,
  go_type,
  plot_type = "dotplot",
  numTerms = 5
)
}
\arguments{
\item{go_ora}{An enrichResult object}

\item{reducedTerms}{A data frame from go_reduce()}

\item{go_type}{The Gene Ontology type}

\item{plot_type}{Either "treemap" or "dotplot"}

\item{numTerms}{The number of GO terms to display}
}
\value{
A ggplot2 object
}
\description{
Creates a plot of the reduced GO terms
}
